requirements.txt
py-altdns
dnsgen
argparse
dnspython
requests
colorama==0.3.7
colored
argparse
requests
datetime
ConfigParser
secure-smtplib
netaddr